feat(#528): move all writers under CustomClassWriter class
